blockchain/lib/requirements.txt
fastecdsa=2.2.3
pycryptodome=3.15.0